+++ ijkmedia/ijkplayer/ijkavformat/ijkiourlhook.c
@@ -20,6 +20,10 @@
  */
 
 #include <assert.h>
+#include <inttypes.h>
+#include <stdio.h>
+#include <stdlib.h>
+#include <string.h>
 #include "ijkiourl.h"
 #include "ijkioprotocol.h"
 #include "ijkplayer/ijkavutil/ijkutils.h"