+++ ijkmedia/ijkplayer/ijkavutil/ijkdict.c
@@ -161,25 +161,25 @@ int ijk_av_dict_set_intptr(IjkAVDictionary **pm, const char *key, uintptr_t valu
 }
 
 uintptr_t ijk_av_dict_strtoptr(char * value) {
-    uintptr_t ptr = NULL;
+    uintptr_t ptr = 0;
     char *next = NULL;
     if(value[0] !='0' || (value[1]|0x20)!='x') {
-        return NULL;
+        return 0;
     }
     ptr = strtoll(value, &next, 16);
     if (next == value) {
-        return NULL;
+        return 0;
     }
     return ptr;
 }
 
 uintptr_t ijk_av_dict_get_intptr(const IjkAVDictionary *m, const char* key) {
-    uintptr_t ptr = NULL;
+    uintptr_t ptr = 0;
     IjkAVDictionaryEntry *t = NULL;
-    if ((t = av_dict_get(m, key, NULL, 0))) {
+    if ((t = ijk_av_dict_get(m, key, NULL, 0))) {
         return ijk_av_dict_strtoptr(t->value);
     }
-    return NULL;
+    return 0;
 }
 
 void ijk_av_dict_free(IjkAVDictionary **pm)